rewrite data seq num handling - 7ab6fa95271
git-svn-id: https://svn.mi.fu-berlin.de/des-testbed/Software/DES-SERT/des-aodv/trunk@7774 ec5dfc5b-5239-0410-a031-e37cc464ff32

+++ src/database/data_seq/data_seq.h
@@ -26,8 +26,8 @@ For further information and questions please use the web site
 
 #include <linux/if_ether.h>
 
-u_int32_t data_get_nextseq(u_int8_t src_addr[ETH_ALEN], u_int8_t dest_addr[ETH_ALEN]);
+u_int16_t data_get_nextseq(u_int8_t src_addr[ETH_ALEN], u_int8_t dest_addr[ETH_ALEN]);
 
-void data_set_seq(u_int8_t src_addr[ETH_ALEN], u_int8_t dest_addr[ETH_ALEN], u_int32_t seq_num);
+void data_set_seq(u_int8_t src_addr[ETH_ALEN], u_int8_t dest_addr[ETH_ALEN], u_int16_t seq_num);
 
 #endif
